chore: rename navbar context folder 2/2
