+++ src/layouts/MainLayout.tsx
@@ -4,14 +4,14 @@ import { QueryParamProvider } from 'use-query-params';
 import { ReactRouter6Adapter } from 'use-query-params/adapters/react-router-6';
 import ModalProvider from '../context/modal/ModalProvider';
 import Header from './components/Header';
-import NavbarNotificationRenderer from '../context/navbarNotifications/NavbarNotificationRenderer';
+import NavbarNotificationRenderer from '../context/navbarNotification/NavbarNotificationRenderer';
 import { Outlet, useNavigation } from 'react-router';
 import Footer from './components/Footer';
 import { Suspense, useCallback, useEffect, useRef } from 'react';
 import noise3D from '../utils/noise/noise3D';
 import { useWindowSize } from 'usehooks-ts';
 import useResizeObserver from '@react-hook/resize-observer';
-import useNavbarNotification from '../context/navbarNotifications/useNavbarNotification';
+import useNavbarNotification from '../context/navbarNotification/useNavbarNotification';
 import { useApp } from '../context/app/useApp';
 import GlobalSpinner from '../components/ui/GlobalSpinner';
 import MenuContextProvider from '../context/menu/MenuContextProvider';
